Fix multi component bug

+++ src/power_grid_model_io/converters/tabular_converter.py
@@ -290,7 +290,7 @@ class TabularConverter(BaseConverter[TabularData]):
 
             # If there are numtiple arrays, concatenate them
             elif len(data_set) > 1:
-                merged[component_name] = np.concatenate(data_set)
+                merged[component_name] = np.concatenate(data_set, dtype=data_set[0].dtype)
 
         return merged
